# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  routes.PublicUserProfile:
    properties:
      bannerUrl:
        type: string
      id:
        type: string
      nickname:
        type: string
    type: object
...
  routes.UpdateUserProfileRequest:
    properties:
      bannerUrl:
        type: string
      nickname:
        type: string
      password:
        type: string
    type: object
  routes.UserProfile:
    properties:
      bannerUrl:
        type: string
      commentReputation:
        type: integer
      email:
        type: string
      id:
        type: string
      nickname:
        type: string
      postReputation:
        type: integer
      registeredAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Reddit-like REST API.
  title: Chirp API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      description: Возвращает публичный профиль пользователя по id
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.PublicUserProfile'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить публичный профиль пользователя
      tags:
      - users
  /users/me:
    get:
      description: Возвращает приватный профиль текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.UserProfile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить свой профиль
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет профиль текущего пользователя
      parameters:
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.UserProfile'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить свой профиль
      tags:
      - users
swagger: "2.0"